+++ src/Modules/OrchardCore.Commerce/Extensions/PriceProviderExtensions.cs
@@ -4,15 +4,18 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace OrchardCore.Commerce.Helpers;
+namespace OrchardCore.Commerce.Extensions;
+
 public static class PriceProviderExtensions
 {
     public static async Task<IDictionary<string, ProductPart>> GetSkuProductsAsync(
         this IProductService productService,
         IList<ShoppingCartItem> items) =>
             (await productService
-                .GetProductsAsync(items.Select(item => item.ProductSku)
-                    .Distinct()
-                    .ToArray()))
+                .GetProductsAsync(items.Select(item => item.ProductSku)))
+                .Distinct()
                 .ToDictionary(productPart => productPart.Sku);
+
+    public static async Task<ShoppingCartItem> AddPriceAsync(this IPriceService priceService, ShoppingCartItem item) =>
+        (await priceService.AddPricesAsync(new[] { item })).Single();
 }
